calc.py
def add2(*args):
    '''
    The 'add2' function itself. It takes an unlimited number of arguments, converts them to their
    appropriate types using the 'conv' function, and adds them together. If any argument is a string,
    it ensures all arguments are strings before concatenating them.

    Parameters:
    *args (int, float, str): The values to be added.

    Returns:
    int, float, str: The result of the addition or concatenation.
    '''
    def conv(value):
        '''
        A helper function to convert values to their appropriate types. The implementation of this
        function is assumed to be provided elsewhere in the calc.py module.
        
        Parameters:
        value (int, float, str): The value to be converted.

        Returns:
        int, float, str: The converted value.
        '''
        # Example conversion logic (this should be replaced with the actual 'conv' function logic)
        try:
            return int(value)
        except ValueError:
            try:
                return float(value)
            except ValueError:
                return str(value)

    if not args:
        return 0

    result = None

    for arg in args:
        arg_conv = conv(arg)
        if result is None:
            result = arg_conv
        else:
            if isinstance(result, str) or isinstance(arg_conv, str):
                result = str(result) + str(arg_conv)
            else:
                result += arg_conv

    return result
